requirements.txt
python-telegram-bot[job-queue]>=21.0
asyncpg>=0.29.0
httpx>=0.27.0
